Refactor, combine Chat and GameChat components
