+++ src/components/Chat.js
@@ -1,26 +1,27 @@
 import { useEffect, useRef, useState, useMemo, useContext, memo } from "react";
 
 import Typography from "@material-ui/core/Typography";
-import { makeStyles } from "@material-ui/core/styles";
 import Tooltip from "@material-ui/core/Tooltip";
 import MoreVertIcon from "@material-ui/icons/MoreVert";
 import MenuItem from "@material-ui/core/MenuItem";
 import Menu from "@material-ui/core/Menu";
+import { makeStyles } from "@material-ui/core/styles";
 
 import User from "./User";
 import InternalLink from "./InternalLink";
 import SimpleInput from "./SimpleInput";
 import Subheading from "./Subheading";
 import Scrollbox from "./Scrollbox";
+import ChatCards from "./ChatCards";
+import ElapsedTime from "./ElapsedTime";
 import firebase from "../firebase";
 import { filter } from "../util";
 import autoscroll from "../utils/autoscroll";
 import useFirebaseQuery from "../hooks/useFirebaseQuery";
 import useStorage from "../hooks/useStorage";
 import { UserContext } from "../context";
-import ElapsedTime from "./ElapsedTime";
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles({
   chatPanel: {
     display: "flex",
     flexDirection: "column",
@@ -51,13 +52,20 @@ const useStyles = makeStyles((theme) => ({
       visibility: "visible",
     },
   },
-}));
+});
 
-function Chat() {
+/** A chat sidebar element, opens lobby chat when the `gameId` prop is not set. */
+function Chat({
+  title,
+  messageLimit,
+  showMessageTimes,
+  gameId,
+  history,
+  gameMode,
+  startedAt,
+}) {
   const user = useContext(UserContext);
   const classes = useStyles();
-
-  const [menuOpenIdx, setMenuOpenIdx] = useState(null);
 
   const chatEl = useRef();
   useEffect(() => {
@@ -65,12 +73,18 @@ function Chat() {
   }, []);
 
   const [input, setInput] = useState("");
+  const [menuOpenIdx, setMenuOpenIdx] = useState(null);
   const [chatHidden, setChatHidden] = useStorage("chat-hidden", "no");
 
+  const databasePath = gameId ? `chats/${gameId}` : "lobbyChat";
   const messagesQuery = useMemo(
     () =>
-      firebase.database().ref("lobbyChat").orderByChild("time").limitToLast(30),
-    []
+      firebase
+        .database()
+        .ref(databasePath)
+        .orderByChild("time")
+        .limitToLast(messageLimit),
+    [databasePath, messageLimit]
   );
   const messages = useFirebaseQuery(messagesQuery);
 
@@ -82,7 +96,7 @@ function Chat() {
           "We detected that your message contains profane language. If you think this was a mistake, please let us know!"
         );
       } else {
-        firebase.database().ref(`lobbyChat`).push({
+        firebase.database().ref(databasePath).push({
           user: user.id,
           message: input,
           time: firebase.database.ServerValue.TIMESTAMP,
@@ -104,14 +118,14 @@ function Chat() {
   };
 
   const handleDelete = (key) => {
-    firebase.database().ref("lobbyChat").child(key).remove();
+    firebase.database().ref(databasePath).child(key).remove();
     handleClose();
   };
 
   const handleDeleteAll = async (uid) => {
     const messages = await firebase
       .database()
-      .ref("lobbyChat")
+      .ref(databasePath)
       .orderByChild("user")
       .equalTo(uid)
       .once("value");
@@ -119,7 +133,7 @@ function Chat() {
     messages.forEach((snap) => {
       updates[snap.key] = null;
     });
-    firebase.database().ref("lobbyChat").update(updates);
+    firebase.database().ref(databasePath).update(updates);
     handleClose();
   };
 
@@ -128,63 +142,88 @@ function Chat() {
     setMenuOpenIdx(null);
   };
 
+  const items = messages;
+  if (gameId && history) {
+    for (let i = 0; i < history.length; i++) {
+      items[`card@${i}`] = history[i];
+    }
+  }
+
+  const timeTooltip = (time, elem) => {
+    if (!showMessageTimes) {
+      return elem;
+    }
+    return (
+      <Tooltip arrow placement="left" title={<ElapsedTime value={time} />}>
+        {elem}
+      </Tooltip>
+    );
+  };
+
   return (
     <section
       className={classes.chatPanel}
       style={{ flexGrow: 1, overflowY: "hidden" }}
     >
       <Subheading className={classes.chatHeader} onClick={toggleChat}>
-        Lobby Chat {chatHidden === "yes" && "(Hidden)"}
+        {title} {chatHidden === "yes" && "(Hidden)"}
       </Subheading>
       <Scrollbox className={classes.chat} ref={chatEl}>
-        {chatHidden !== "yes" &&
-          Object.entries(messages)
-            .sort((a, b) => a[1].time - b[1].time)
-            .map(([key, msg]) => (
-              <div
+        {Object.entries(items)
+          .sort(([, a], [, b]) => a.time - b.time)
+          .map(([key, item]) =>
+            key.startsWith("card@") ? (
+              <ChatCards
                 key={key}
-                style={{ display: "flex", flexDirection: "row" }}
-                className={classes.message}
-              >
-                <Tooltip
-                  arrow
-                  placement="left"
-                  title={<ElapsedTime value={msg.time} />}
+                item={item}
+                gameMode={gameMode}
+                startedAt={startedAt}
+              />
+            ) : (
+              chatHidden !== "yes" && (
+                <div
+                  key={key}
+                  style={{ display: "flex", flexDirection: "row" }}
+                  className={classes.message}
                 >
-                  <Typography variant="body2" gutterBottom>
-                    <User
-                      id={msg.user}
-                      component={InternalLink}
-                      to={`/profile/${msg.user}`}
-                      underline="none"
+                  {timeTooltip(
+                    item.time,
+                    <Typography variant="body2" gutterBottom>
+                      <User
+                        id={item.user}
+                        component={InternalLink}
+                        to={`/profile/${item.user}`}
+                        underline="none"
+                      />
+                      : {item.message}
+                    </Typography>
+                  )}
+                  {user.admin && (
+                    <MoreVertIcon
+                      aria-controls="admin-menu"
+                      color="inherit"
+                      className={classes.vertIcon}
+                      onClick={(e) => handleClickVertIcon(e, key)}
                     />
-                    : {msg.message}
-                  </Typography>
-                </Tooltip>
-                {user.admin && (
-                  <MoreVertIcon
-                    aria-controls="admin-menu"
-                    color="inherit"
-                    className={classes.vertIcon}
-                    onClick={(e) => handleClickVertIcon(e, key)}
-                  />
-                )}
+                  )}
 
-                <Menu
-                  id="admin-menu"
-                  anchorEl={anchorEl}
-                  open={Boolean(anchorEl) && key === menuOpenIdx}
-                  onClose={handleClose}
-                >
-                  <MenuItem onClick={() => handleDelete(key)}>
-                    Delete message
-                  </MenuItem>
-                  <MenuItem onClick={() => handleDeleteAll(msg.user)}>
-                    Delete all from user
-                  </MenuItem>
-                </Menu>
-              </div>
-            ))}
+                  <Menu
+                    id="admin-menu"
+                    anchorEl={anchorEl}
+                    open={Boolean(anchorEl) && key === menuOpenIdx}
+                    onClose={handleClose}
+                  >
+                    <MenuItem onClick={() => handleDelete(key)}>
+                      Delete message
+                    </MenuItem>
+                    <MenuItem onClick={() => handleDeleteAll(item.user)}>
+                      Delete all from user
+                    </MenuItem>
+                  </Menu>
+                </div>
+              )
+            )
+          )}
       </Scrollbox>
       <form onSubmit={handleSubmit}>
         <SimpleInput
